+++ packages/ui/src/views/user-management/components/dialogs/edit-user/schemas.ts
@@ -1,0 +1,7 @@
+import { z } from 'zod'
+
+export const newUserSchema = z.object({
+  userID: z.string(),
+  email: z.string().email({ message: 'Please provide a valid email, ex: example@yourcompany.com' }),
+  displayName: z.string().min(1, { message: 'Please provide a display name' })
+})
