+++ packages/ui/src/views/user-management/components/dialogs/edit-user/types.ts
@@ -1,4 +1,8 @@
-import { IAdminListUsersStore } from 'dist/views'
+import { IAdminListUsersStore } from '@/views'
+import { newUserSchema } from '@/views/user-management/components/dialogs/edit-user/schemas'
+import { z } from 'zod'
+
+export type MemberFields = z.infer<typeof newUserSchema>
 
 export interface IEditUserDialogProps {
   isSubmitting: boolean
