+++ packages/ui/src/views/user-management/components/dialogs/reset-password/reset-password.tsx
@@ -1,8 +1,7 @@
 import { FC } from 'react'
 
 import { AlertDialog, Button, CopyButton, Input, Text } from '@/components'
-
-import { IResetPasswordDialogProps } from './types'
+import { IResetPasswordDialogProps } from '@/views/user-management/components/dialogs/reset-password/types'
 
 export const ResetPasswordDialog: FC<IResetPasswordDialogProps> = ({
   open,
