+++ packages/ui/src/views/user-management/components/dialogs/reset-password/types.ts
@@ -1,4 +1,4 @@
-import { IAdminListUsersStore } from 'dist/views'
+import { IAdminListUsersStore } from '@/views'
 
 export interface IResetPasswordDialogProps {
   onClose: () => void
