+++ packages/ui/src/views/user-management/components/page-components/content/components/users-list/index.ts
@@ -1,0 +1,1 @@
+export { UsersList } from './users-list'
