+++ packages/ui/src/views/user-management/components/page-components/content/components/users-list/types.ts
@@ -1,0 +1,6 @@
+import { UsersProps } from '@views/user-management/types'
+
+export interface PageProps {
+  users: UsersProps[]
+  handleDialogOpen: (user: UsersProps | null, dialogLabel: string) => void
+}
